Python Code/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.33.3
idna==3.3
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
torch==1.11.0
torchvision==0.12.0
typing-extensions==4.2.0
urllib3==1.26.9
